requirement.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.1
django-allauth==0.44.0
django-heroku==0.3.1
django-jsonfield==1.4.1
django-truncate==0.1
gunicorn==20.1.0
heroku==0.1.4
oauthlib==3.1.0
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
PyJWT==2.1.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
whitenoise==5.2.0
